MODDCB-61 Implementing post api in borrowing library
