+++ src/main/java/org/folio/dcb/service/impl/BorrowingLibraryServiceImpl.java
@@ -21,10 +21,12 @@ public class BorrowingLibraryServiceImpl implements LibraryService {
 
   private final CirculationService circulationService;
   private final TransactionRepository transactionRepository;
+  private final LibraryUtil libraryUtil;
 
   @Override
   public TransactionStatusResponse createCirculation(String dcbTransactionId, DcbTransaction dcbTransaction, String pickupServicePointId) {
-    return null;
+    return libraryUtil.
+      createBorrowingLibraryTransaction(dcbTransactionId, dcbTransaction, pickupServicePointId);
   }
 
   @Override
