+++ src/main/java/org/folio/dcb/service/impl/LibraryUtil.java
@@ -1,0 +1,53 @@
+package org.folio.dcb.service.impl;
+
+import lombok.RequiredArgsConstructor;
+import org.folio.dcb.domain.dto.DcbTransaction;
+import org.folio.dcb.domain.dto.TransactionStatusResponse;
+import org.folio.dcb.domain.entity.TransactionEntity;
+import org.folio.dcb.domain.mapper.TransactionMapper;
+import org.folio.dcb.repository.TransactionRepository;
+import org.folio.dcb.service.CirculationItemService;
+import org.folio.dcb.service.CirculationService;
+import org.folio.dcb.service.RequestService;
+import org.folio.dcb.service.UserService;
+import org.springframework.stereotype.Component;
+
+import java.util.Objects;
+
+import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CREATED;
+
+@Component
+@RequiredArgsConstructor
+public class LibraryUtil {
+
+  private final UserService userService;
+  private final RequestService requestService;
+  private final CirculationItemService circulationItemService;
+  private final TransactionRepository transactionRepository;
+  private final TransactionMapper transactionMapper;
+
+  public TransactionStatusResponse createBorrowingLibraryTransaction(String dcbTransactionId, DcbTransaction dcbTransaction, String pickupServicePointId) {
+    var itemVirtual = dcbTransaction.getItem();
+    var patron = dcbTransaction.getPatron();
+
+    var user = userService.fetchUser(patron); //user is needed, but shouldn't be generated. it should be fetched.
+    circulationItemService.checkIfItemExistsAndCreate(itemVirtual, pickupServicePointId);
+
+    requestService.createHoldItemRequest(user, itemVirtual, pickupServicePointId);
+    saveDcbTransaction(dcbTransactionId, dcbTransaction, null);
+
+    return TransactionStatusResponse.builder()
+      .status(TransactionStatusResponse.StatusEnum.CREATED)
+      .item(itemVirtual)
+      .patron(patron)
+      .build();
+  }
+  public void saveDcbTransaction(String dcbTransactionId, DcbTransaction dcbTransaction, String requestId) {
+    TransactionEntity transactionEntity = transactionMapper.mapToEntity(dcbTransactionId, dcbTransaction);
+    if (Objects.isNull(transactionEntity)) {
+      throw new IllegalArgumentException("Transaction Entity is null");
+    }
+    transactionEntity.setStatus(CREATED);
+    transactionRepository.save(transactionEntity);
+  }
+}
