+++ src/main/java/org/folio/dcb/service/impl/TransactionsServiceImpl.java
@@ -44,17 +44,12 @@ public class TransactionsServiceImpl implements TransactionsService {
     checkTransactionExistsAndThrow(dcbTransactionId);
     ServicePointRequest pickupServicePoint = servicePointService.createServicePoint(dcbTransaction.getPickup());
 
-    TransactionStatusResponse circulationStatusResponse =
-      switch (dcbTransaction.getRole()) {
-        case LENDER -> lendingLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
-        case BORROWING_PICKUP -> borrowingPickupLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
-        case PICKUP -> pickupLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
-        default -> throw new IllegalArgumentException("Other roles are not implemented");
-      };
-
-    saveDcbTransaction(dcbTransactionId, dcbTransaction);
-
-    return circulationStatusResponse;
+    return switch (dcbTransaction.getRole()) {
+      case LENDER -> lendingLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
+      case BORROWING_PICKUP -> borrowingPickupLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
+      case PICKUP -> pickupLibraryService.createCirculation(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
+      default -> throw new IllegalArgumentException("Other roles are not implemented");
+    };
   }
 
   @Override
@@ -106,14 +101,5 @@ public class TransactionsServiceImpl implements TransactionsService {
       throw new ResourceAlreadyExistException(
         String.format("unable to create transaction with id %s as it already exists", dcbTransactionId));
     }
-  }
-
-  private void saveDcbTransaction(String dcbTransactionId, DcbTransaction dcbTransaction) {
-    TransactionEntity transactionEntity = transactionMapper.mapToEntity(dcbTransactionId, dcbTransaction);
-    if (Objects.isNull(transactionEntity)) {
-      throw new IllegalArgumentException("Transaction Entity is null");
-    }
-    transactionEntity.setStatus(CREATED);
-    transactionRepository.save(transactionEntity);
   }
 }
